Add A Folder Structure To The Project Directory
* This Commit adds a folder structure to the project.
* A hooks folder holds all hooks in the project.
* The commponenets folder holds components used in the project.
* A pages folder holds all pages in the project.
* A utils folder holds all helper functions in the project.

Author: Hamza Kyamanywa <https://github.com/hsanshine>
