+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import "./App.css";
-import { BionicReaderPage } from "./BionicReaderPage";
+import { BionicReaderPage } from "./pages/BionicReaderPage";
 
 function App() {
   const [count, setCount] = useState(0);
