+++ src/pages/BionicReaderPage.tsx
@@ -1,81 +1,78 @@
 import React, { FC, useState } from "react";
 
-//helper
-const timeout = (ms: number) => {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-};
+import { useTextProcessing } from "../hooks/useTextProcessing";
 
 //useHook
-const useTextProcessing = () => {
-  const [isDisabled, setIsDisabled] = useState(false);
-  const [text, setText] = useState("");
-  const [pretext, setPretext] = useState("");
-  const [listPrepText, setListPrepText] = useState([] as JSX.Element[]);
+// const useTextProcessing = () => {
+//   const [isDisabled, setIsDisabled] = useState(false);
+//   const [text, setText] = useState("");
+//   const [pretext, setPretext] = useState("");
+//   const [listPrepText, setListPrepText] = useState([] as JSX.Element[]);
 
-  const onChangeTextare = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setText(e?.target.value);
-  };
+//   const onChangeTextare = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+//     setText(e?.target.value);
+//   };
 
-  const processData = () => {
-    const prepText = text.split(" ");
-    console.log(prepText);
+//   const processData = () => {
+//     const prepText = text.split(" ");
+//     console.log(prepText);
 
-    const listText = prepText.map((elem) => {
-      let preElem = elem;
-      let showNewLine = false;
-      const match = /\r|\n/.exec(elem);
-      if (match) {
-        preElem = elem.trim();
-        showNewLine = true;
-      }
+//     const listText = prepText.map((elem) => {
+//       let preElem = elem;
+//       let showNewLine = false;
+//       const match = /\r|\n/.exec(elem);
+//       if (match) {
+//         preElem = elem.trim();
+//         showNewLine = true;
+//       }
 
-      const mid = Math.floor(preElem.length / 2);
+//       const mid = Math.floor(preElem.length / 2);
 
-      console.log(preElem.slice(mid));
-      return (
-        <>
-          {showNewLine && (
-            <>
-              <br />
-              <br />
-            </>
-          )}
-          <span className="bio-letter">{preElem.slice(0, mid)}</span>
-          {preElem.slice(mid)}
-        </>
-      );
-    });
+//       console.log(preElem.slice(mid));
+//       return (
+//         <>
+//           {showNewLine && (
+//             <>
+//               <br />
+//               <br />
+//             </>
+//           )}
+//           <span className="bio-letter">{preElem.slice(0, mid)}</span>
+//           {preElem.slice(mid)}
+//         </>
+//       );
+//     });
 
-    return listText;
-  };
+//     return listText;
+//   };
 
-  const onClickButton = async (e: React.MouseEvent<HTMLButtonElement>) => {
-    setPretext("processing...");
-    setListPrepText([]);
+//   const onClickButton = async (e: React.MouseEvent<HTMLButtonElement>) => {
+//     setPretext("processing...");
+//     setListPrepText([]);
 
-    console.log("wait...");
-    setIsDisabled(true);
+//     console.log("wait...");
+//     setIsDisabled(true);
 
-    const listPrepText = processData();
-    setPretext("processing...");
+//     const listPrepText = processData();
+//     setPretext("processing...");
 
-    await timeout(2000);
+//     await timeout(2000);
 
-    setIsDisabled(false);
-    setListPrepText(listPrepText);
-    setPretext("");
-    console.log("done...");
-  };
+//     setIsDisabled(false);
+//     setListPrepText(listPrepText);
+//     setPretext("");
+//     console.log("done...");
+//   };
 
-  return {
-    listPrepText,
-    pretext,
-    isDisabled,
-    onClickButton,
-    processData,
-    onChangeTextare,
-  };
-};
+//   return {
+//     listPrepText,
+//     pretext,
+//     isDisabled,
+//     onClickButton,
+//     processData,
+//     onChangeTextare,
+//   };
+// };
 
 export const BionicReaderPage: FC = () => {
   const { listPrepText, isDisabled, onClickButton, onChangeTextare, pretext } =
