+++ src/util/helpers.tsx
@@ -1,0 +1,3 @@
+export const timeout = (ms: number) => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+};
